CPC: moved AttrDict -> cpcbase
